requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==4.2.16
django-cors-headers==4.6.0
django-password-validators==1.7.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
Faker==30.8.2
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.3
